requirements.txt
git+https://github.com/huggingface/optimum-graphcore.git
gradio 
pillow
diffusers[torch]==0.9